pyproject.toml
[tool.poetry]
name = "lab2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lmql = "^0.7.3"
poetry-dotenv-plugin = "^0.2.0"
numpy = "1.25.2"
rich = "^13.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
